McMaster Attitudes Data Set.

A data set containing information about students commuting to McMaster University. Data were collected by means of a travel survey in the fall of 2010. Respondents were asked about their mode of travel to McMaster University, in Hamilton, Canada. They were also asked to respond to a series of attitudinal statements. The data set contains relevant attributes about the respondents. The format of the table is wide, with each row representing an individual respondent. These variables were augmented with information from the 2011 Census of Canada and land use information from local sources.

Usage

data(mc_attitudes)

Format

A data frame with 1230 rows and 39 variables:

id

Unique identifier for respondents

choice

Mode of transportation chosen: Car, HSR (local transit), Walk

LAT

Latitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

LONG

Longitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

license

Indicator variable for holding a driver license: Yes/No

vehicle

Indicator variable for having individual access to a private car (i.e., car not shared with other household members): Yes/No

gender

Woman/Man

age

Age in years

visa

Visa status of student: Domestic

living_arrangements

Living arrangements DURING THE SCHOOL YEAR are: 1:Living with my family or relatives/By myself off-campus/Shared off-campus accommodations

level

Level of studies: Undergraduate Year I (UG I)/Undergraduate Year II (UG I)/Undergraduate Year III (UG III)/Undergraduate Year IV (UG IV)/Undergraduate Year V (UG V)/Masters/PhD/Other (Diploma or Certificate)

Active_Neighborhood

Response to statement "I like to live in a neighborhood where there's a lot going on": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

Community

Response to statement "There is a sense of community in my neighborhood": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

Neighbors

Response to statement "I know my neighbors well": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

Safe_Walk

Response to statement "I feel safe and secure when walking in my neighborhood": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

Shops_Important

Response to statement "Having shops and services within walking distance of my home is important to me": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

Travel_Alone

Response to statement "I like traveling alone": Five point likert scale from Strongly Disagree to Strongly Agree, including a neutral point

DAUID

Unique ID of Dissemination Area of the Census

Rate_Couple_Child

Proportion of families that are couples with children out of all families in Dissemination Area

Rate_SW_Child

Proportion of families that are single parent (woman) with children out of all families in Dissemination Area

Rate_SM_Child

Proportion of families that are single parent (man) with children out of all families in Dissemination Area

Mean_Children

Average number of children at home per census family in Dissemination Area

Rate_Non_Canadian

Proportion of non-Canadian residents to total population in Dissemination Area

Median_HH_Income

Median family income in Dissemination Area in $10,000

Mean_HH_Income

Mean family income in Dissemination Area in $10,000

Rate_Unemployment

Unemployment rate in Dissemination Area

Rate_1yr_Move

Proportion of residents that moved to Dissemination Area in the year previous to the census

Rate_5yr_Move

Proportion of residents that moved to Dissemination Area in the 5 years previous to the census

Rate_Public

Proportion of residents in Dissemination Area that used public transportation to commute to work

Rate_Walk

Proportion of residents in Dissemination Area that walked to work

Rate_Cycle

Proportion of residents in Dissemination Area that cycled to work

AREA

Area of Dissemination Area in square kilometers

LUM

Entropy-based land use mix index in Dissemination Area: lower values mean more homogenous land uses, higher values greater mix of uses

SIDEWALK_DENSITY

Sidewalk density in Dissemination Area in km/sq.km

STREET_DENSITY

Street density in Dissemination Area in km/sq.km

INTERSECTION_DENSITY

Intersection density in Dissemination Area in 1/sq.km

SF_P_RATIO

Ratio of building footprint to parcel area in Dissemination Area: a measure of density of development

EMPLOYMENT_DENSITY

Employment density in Dissemination Area in jobs/sq.km

POPULATION_DENSITY

Population density in Dissemination Area in people/sq.km

Source

P\'aez, A. (2013). Mapping travelers’ attitudes: does space matter?. Journal of Transport Geography, 26, 117-125. (https://doi.org/10.1016/j.jtrangeo.2012.09.002)

Examples

 data(mc_attitudes)
 summary(mc_attitudes$Community)
#> STRONGLY DISAGREE          DISAGREE           NEUTRAL             AGREE 
#>                69               299               414               391 
#>    STRONGLY AGREE 
#>                57