McMaster Commuting Dataset (csv file).

A delimited text file containing information about students commuting to McMaster University. Data were collected by means of a travel survey in the fall of 2010. Respondents were asked about their mode of travel to McMaster University, in Hamilton, Canada. They were also asked about the modes available to them. The characteristics of the trips were self-reported or imputed. The dataset also contains relevant attributes about the respondents. The format of the table is long, with each row representing a choice situation.

Usage

data(mc_commute_long)

Format

An comma separated text file :

RespondentID

Unique identifier for respondents

choice

numeric variable indicating modes of transportation: (1) Cycle, (2) Walk, (3) HSR (local transit), (4) Car

avcycle

Indicator variable for availability of cycling: (1) Yes, (0) No

avwalk

Indicator variable for availability of walking: (1) Yes, (0) No

avhsr

Indicator variable for availability of HSR: (1) Yes, (0) No

avcar

Indicator variable for availability of car: (1) Yes, (0) No

timecycle

Travel time by cycling in minutes (when mode not available coded as 100000)

timewalk

Travel time by walking in minutes (when mode not available coded as 100000)

accesshsr

Access time to HSR in minutes

waitingtimehsr

Waiting time when travelling by HSR in minutes

transfer

Number of transit transfers

timecar

Travel time by car in minutes (when mode not available coded as 100000)

parking

Indicator variable for holding a university parking permit: (1) Yes, (0) No

vehind

Indicator variable for having individual access to a private car (i.e., car not shared with other household members): (1) Yes, (0) No

owncycle

Indicator variable for owning a bicycle: (1) Yes, (0) No

gender

Indicator variable for gender: (1) Woman, (0) Man

age

Age in years

solo

Indicator variable for living in solo accommodations: (1) Yes, (0) No

shared

Indicator variable for a student living in a rental property with other students, i.e., shared accommodations: (1) Yes, (0) No

family

Indicator variable for living in family home: (1) Yes, (0) No

child

Indicator variable for presence of dependent minors in household: (1) Yes, (0) No

primary_caregiver

Indicator variable for being the primary caregiver of dependent minors in household: (1) Yes, (0) No, (100000) No dependent minors in household

LAT

Latitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

LONG

Longitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

DAUID

Unique ID of the dissemination area of place of residence according to the 2006 census of Canada

mhi

Median household income of the dissemination area of place of residence according to the 2006 census of Canada

dwell_den

Dwelling density in the dissemination area of the place of residence in $1/km^2$

lum

Entropy of the land use mix

st_den

Street density in the dissemination area of the place of residence in $km/km^2$

inter_den

Intersection density in the dissemination area of the place of residence in $1/km^2$

SF_P_ratio

Ratio of surface of plots to built area

side_density

Sidewalk density in the dissemination area of the place of residence in $km/km^2$

Shelters_SD

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Disagree, 0 otherwise

Shelters_D

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Disagree, 0 otherwise

Shelters_A

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Agree, 0 otherwise

Shelters_SA

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Agree, 0 otherwise

Source

Whalen, K. E., P\'aez, A., & Carrasco, J. A. (2013). Mode choice of university students commuting to school and the role of active travel. Journal of Transport Geography, 31, 132-142. (https://doi.org/10.1016/j.jtrangeo.2013.06.008)

Examples

 library(readr)
 mc_commute <- read_csv(system.file("extdata", "mc_commute.csv", package = "discrtr"))
#> Rows: 1376 Columns: 39
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> dbl (39): RespondentID, choice, avcycle, avwalk, avhsr, avcar, timecycle, ti...
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.