McMaster Commuting Dataset (long format).

A dataset containing information about commuting to McMaster University. Data were collected by means of a travel survey in the fall of 2010. Respondents were asked about their mode of travel to McMaster University, in Hamilton, Canada. They were also asked about the modes available to them. The characteristics of the trips were self-reported or imputed. The dataset also contains relevant attributes about the respondents. The format of the table is long, with each row representing a choice situation.

Usage

data(mc_commute_long)

Format

An indexed data frame of class dfidx with 5500 rows and 62 variables:

id

Unique identifier for respondents

choice

Mode of transportation chosen: Cycle, Walk, HSR (local transit), Car

parking

Indicator variable for holding a university parking permit: Yes/No

vehind

Indicator variable for having individual access to a private car (i.e., car not shared with other household members): Yes/No

gender

Woman/Man

age

Age in years

shared

Indicator variable for Indicator variable for a student living in a rental property with other students, i.e., living in shared accommodations

family

Indicator variable for living in family home

child

Indicator variable for presence of dependent minors in household: Yes/No

street_density

Street density in the dissemination area of the place of residence in $km/km^2$

sidewalk_density

Sidewalk density in the dissemination area of the place of residence in $km/km^2$

LAT

Latitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

LONG

Longitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

PersonalVehComf_SD

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Strongly Disagree, 0 otherwise

PersonalVehComf_D

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Disagree, 0 otherwise

PersonalVehComf_A

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Agree, 0 otherwise

PersonalVehComf_SA

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Strongly Agree, 0 otherwise

Fun_SD

Response to statement "Getting there is half the fun": 1 if Strongly Disagree, 0 otherwise

Fun_D

Response to statement "Getting there is half the fun": 1 if Disagree, 0 otherwise

Fun_A

Response to statement "Getting there is half the fun": 1 if Agree, 0 otherwise

Fun_SA

Response to statement "Getting there is half the fun": 1 if Strongly Agree, 0 otherwise

ActiveNeigh_SD

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Strongly Disagree, 0 otherwise

ActiveNeigh_D

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Disagree, 0 otherwise

ActiveNeigh_A

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Agree, 0 otherwise

ActiveNeigh_SA

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Strongly Agree, 0 otherwise

UsefulTrans_SD

Response to statement "My commute trip is a useful transition between home and school": 1 if Strongly Disagree, 0 otherwise

UsefulTrans_D

Response to statement "My commute trip is a useful transition between home and school": 1 if Disagree, 0 otherwise

UsefulTrans_A

Response to statement "My commute trip is a useful transition between home and school": 1 if Agree, 0 otherwise

UsefulTrans_SA

Response to statement "My commute trip is a useful transition between home and school": 1 if Strongly Agree, 0 otherwise

BusComf_SD

Response to statement "The buses I usually travel in are comfortable": 1 if Strongly Disagree, 0 otherwise

BusComf_D

Response to statement "The buses I usually travel in are comfortable": 1 if Disagree, 0 otherwise

BusComf_A

Response to statement "The buses I usually travel in are comfortable": 1 if Agree, 0 otherwise

BusComf_SA

Response to statement "The buses I usually travel in are comfortable": 1 if Strongly Agree, 0 otherwise

TravelAlone_SD

Response to statement "I like traveling alone": 1 if Strongly Disagree, 0 otherwise

TravelAlone_D

Response to statement "I like traveling alone": 1 if Disagree, 0 otherwise

TravelAlone_A

Response to statement "I like traveling alone": 1 if Agree, 0 otherwise

TravelAlone_SA

Response to statement "I like traveling alone": 1 if Strongly Agree, 0 otherwise

Shelters_SD

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Disagree, 0 otherwise

Shelters_D

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Disagree, 0 otherwise

Shelters_A

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Agree, 0 otherwise

Shelters_SA

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Agree, 0 otherwise

Community_SD

Response to statement "There is a sense of community in my neighborhood": 1 if Strongly Disagree, 0 otherwise

Community_D

Response to statement "There is a sense of community in my neighborhood": 1 if Disagree, 0 otherwise

Community_A

Response to statement "There is a sense of community in my neighborhood": 1 if Agree, 0 otherwise

Community_SA

Response to statement "There is a sense of community in my neighborhood": 1 if Agree, 0 otherwise

personal_veh_comfortable

Response to statement "The personal vehicles I usually travel in are comfortable" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

getting_there_fun

Response to statement "Getting there is half the fun" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

like_active_neighborhood

Response to statement "I like to live in a neighborhood where there's a lot going on" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

commute_useful_transition

Response to statement "My commute trip is a useful transition between home and school" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

buses_comfortable

Response to statement "The buses I usually travel in are comfortable" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

prefer_travel_alone

Response to statement "I like traveling alone" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

shelter_good_quality

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

sense_community

Response to statement "There is a sense of community in my neighborhood" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree1 if Strongly Disagree, 0 otherwise

numna

Total number of available alternatives, e.g., if bus and walk = 2, if all four = 4

alt

A factor with the alternatives in the choice set

available

An indicator variable for availability of the alternative

time.Cycle

Travel time by alternative

access

Access time to transit in minutes (i.e., reaching a transit stop)

wait

Waiting time transit (i.e., at a transit stop)

transfer

Number of transit transfers

idx

Index for the table: includes chid (identifier for the choice) and alt (the alternative)

Source

Whalen, K. E., P\'aez, A., & Carrasco, J. A. (2013). Mode choice of university students commuting to school and the role of active travel. Journal of Transport Geography, 31, 132-142. (https://doi.org/10.1016/j.jtrangeo.2013.06.008)

Examples

 data(mc_commute_long)
 library(dfidx)
#> 
#> Attaching package: ‘dfidx’
#> The following object is masked from ‘package:stats’:
#> 
#>     filter
 print(mc_commute_long, n = 12)
#> # A tibble: 5,500 × 62
#>         id choice parking vehind gender   age shared family child street_density
#>  *   <dbl> <lgl>  <fct>   <fct>  <chr>  <dbl> <fct>  <fct>  <fct>          <dbl>
#>  1  5.67e8 FALSE  No      No     Man       21 Livin… No     No              14.4
#>  2  5.67e8 FALSE  No      No     Man       21 Livin… No     No              14.4
#>  3  5.67e8 TRUE   No      No     Man       21 Livin… No     No              14.4
#>  4  5.67e8 FALSE  No      No     Man       21 Livin… No     No              14.4
#>  5  5.67e8 FALSE  No      Yes    Man       23 No     No     No              19.5
#>  6  5.67e8 FALSE  No      Yes    Man       23 No     No     No              19.5
#>  7  5.67e8 TRUE   No      Yes    Man       23 No     No     No              19.5
#>  8  5.67e8 FALSE  No      Yes    Man       23 No     No     No              19.5
#>  9  5.67e8 FALSE  No      No     Man       20 No     Livin… No              13.6
#> 10  5.67e8 FALSE  No      No     Man       20 No     Livin… No              13.6
#> 11  5.67e8 TRUE   No      No     Man       20 No     Livin… No              13.6
#> 12  5.67e8 FALSE  No      No     Man       20 No     Livin… No              13.6
#> # … with 5,488 more rows, and 52 more variables: sidewalk_density <dbl>,
#> #   LAT <dbl>, LONG <dbl>, PersonalVehComf_SD <dbl>, PersonalVehComf_D <dbl>,
#> #   PersonalVehComf_A <dbl>, PersonalVehComf_SA <dbl>, Fun_SD <dbl>,
#> #   Fun_D <dbl>, Fun_A <dbl>, Fun_SA <dbl>, ActiveNeigh_SD <dbl>,
#> #   ActiveNeigh_D <dbl>, ActiveNeigh_A <dbl>, ActiveNeigh_SA <dbl>,
#> #   UsefulTrans_SD <dbl>, UsefulTrans_D <dbl>, UsefulTrans_A <dbl>,
#> #   UsefulTrans_SA <dbl>, BusComf_SD <dbl>, BusComf_D <dbl>, BusComf_A <dbl>, …
#> 
#> ~~~ indexes ~~~~
#>    chid   alt
#> 1     1   Car
#> 2     1 Cycle
#> 3     1   HSR
#> 4     1  Walk
#> 5     2   Car
#> 6     2 Cycle
#> 7     2   HSR
#> 8     2  Walk
#> 9     3   Car
#> 10    3 Cycle
#> 11    3   HSR
#> 12    3  Walk
#> indexes:  1, 2