McMaster Commuting Dataset (wide format).

A dataset containing information about commuting to McMaster University. Data were collected by means of a travel survey in the fall of 2010. Respondents were asked about their mode of travel to McMaster University, in Hamilton, Canada. They were also asked about the modes available to them. The characteristics of the trips were self-reported or imputed. The dataset also contains relevant attributes about the respondents. The format of the table is wide, with each individual decision maker in a row.

Usage

data(mc_commute_wide)

Format

A data frame with 1375 rows and 74 variables:

id

Unique identifier for respondents

choice

Mode of transportation chosen: Cycle, Walk, HSR (local transit), Car

available.Cycle

An indicator variable for availability of cycling: Yes/No

available.Walk

An indicator variable for availability of walking: Yes/No

available.HSR

An indicator variable for availability of buses: Yes/No

available.Car

An indicator variable for availability of car: Yes/No

time.Cycle

Travel time by bicycle to campus in minutes

time.Walk

Travel time by walking to campus in minutes

time.HSR

Travel time by bus to campus in minutes

time.Car

Travel time by car to campus in minutes

access.Cycle

Access time to mode in minutes (i.e., reaching a transit stop)

access.Walk

Access time to mode in minutes (i.e., reaching a transit stop)

access.HSR

Access time to mode in minutes (i.e., reaching a transit stop)

access.Car

Access time to mode in minutes (i.e., reaching a transit stop)

wait.Cycle

Waiting time (i.e., at a transit stop)

wait.Walk

Waiting time (i.e., at a transit stop)

wait.HSR

Waiting time (i.e., at a transit stop)

wait.Car

Waiting time (i.e., at a transit stop)

transfer.Walk

Number of transfers when travelling by transit

transfer.Cycle

Number of transfers when travelling by transit

transfer.HSR

Number of transfers when travelling by transit

transfer.Car

Number of transfers when travelling by transit

parking

Indicator variable for holding a university parking permit: Yes/No

vehind

Indicator variable for having individual access to a private car (i.e., car not shared with other household members): Yes/No

sex

Woman/Man

age

Age in years

shared

Indicator variable for a student living in a rental property with other students, i.e., living in shared accommodations

family

Indicator variable for living in family home

child

Indicator variable for presence of dependent minors in household: Yes/No

street_density

Street density in the dissemination area of the place of residence in $km/km^2$

sidewalk_density

Sidewalk density in the dissemination area of the place of residence in $km/km^2$

LAT

Latitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

LONG

Longitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

PersonalVehComf_SD

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Strongly Disagree, 0 otherwise

PersonalVehComf_D

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Disagree, 0 otherwise

PersonalVehComf_A

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Agree, 0 otherwise

PersonalVehComf_SA

Response to statement "The personal vehicles I usually travel in are comfortable": 1 if Strongly Agree, 0 otherwise

Fun_SD

Response to statement "Getting there is half the fun": 1 if Strongly Disagree, 0 otherwise

Fun_D

Response to statement "Getting there is half the fun": 1 if Disagree, 0 otherwise

Fun_A

Response to statement "Getting there is half the fun": 1 if Agree, 0 otherwise

Fun_SA

Response to statement "Getting there is half the fun": 1 if Strongly Agree, 0 otherwise

ActiveNeigh_SD

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Strongly Disagree, 0 otherwise

ActiveNeigh_D

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Disagree, 0 otherwise

ActiveNeigh_A

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Agree, 0 otherwise

ActiveNeigh_SA

Response to statement "I like to live in a neighborhood where there's a lot going on": 1 if Strongly Agree, 0 otherwise

UsefulTrans_SD

Response to statement "My commute trip is a useful transition between home and school": 1 if Strongly Disagree, 0 otherwise

UsefulTrans_D

Response to statement "My commute trip is a useful transition between home and school": 1 if Disagree, 0 otherwise

UsefulTrans_A

Response to statement "My commute trip is a useful transition between home and school": 1 if Agree, 0 otherwise

UsefulTrans_SA

Response to statement "My commute trip is a useful transition between home and school": 1 if Strongly Agree, 0 otherwise

BusComf_SD

Response to statement "The buses I usually travel in are comfortable": 1 if Strongly Disagree, 0 otherwise

BusComf_D

Response to statement "The buses I usually travel in are comfortable": 1 if Disagree, 0 otherwise

BusComf_A

Response to statement "The buses I usually travel in are comfortable": 1 if Agree, 0 otherwise

BusComf_SA

Response to statement "The buses I usually travel in are comfortable": 1 if Strongly Agree, 0 otherwise

TravelAlone_SD

Response to statement "I like traveling alone": 1 if Strongly Disagree, 0 otherwise

TravelAlone_D

Response to statement "I like traveling alone": 1 if Disagree, 0 otherwise

TravelAlone_A

Response to statement "I like traveling alone": 1 if Agree, 0 otherwise

TravelAlone_SA

Response to statement "I like traveling alone": 1 if Strongly Agree, 0 otherwise

Shelters_SD

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Disagree, 0 otherwise

Shelters_D

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Disagree, 0 otherwise

Shelters_A

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Agree, 0 otherwise

Shelters_SA

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality": 1 if Strongly Agree, 0 otherwise

Community_SD

Response to statement "There is a sense of community in my neighborhood": 1 if Strongly Disagree, 0 otherwise

Community_D

Response to statement "There is a sense of community in my neighborhood": 1 if Disagree, 0 otherwise

Community_A

Response to statement "There is a sense of community in my neighborhood": 1 if Agree, 0 otherwise

Community_SA

Response to statement "There is a sense of community in my neighborhood": 1 if Agree, 0 otherwise

personal_veh_comfortable

Response to statement "The personal vehicles I usually travel in are comfortable" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

getting_there_fun

Response to statement "Getting there is half the fun" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

like_active_neighborhood

Response to statement "I like to live in a neighborhood where there's a lot going on" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

commute_useful_transition

Response to statement "My commute trip is a useful transition between home and school" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

buses_comfortable

Response to statement "The buses I usually travel in are comfortable" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

prefer_travel_alone

Response to statement "I like traveling alone" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

shelter_good_quality

Response to statement "Shelters and other public transportation facilities that I commonly use are of good quality" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree

sense_community

Response to statement "There is a sense of community in my neighborhood" as factor: SD-Strongly Disagree, D-Disagree, N-Neutral, A-Agree, SA-Strongly Agree1 if Strongly Disagree, 0 otherwise

numna

Total number of available alternatives, e.g., if bus and walk = 2, if all four = 4

Source

Whalen, K. E., P\'aez, A., & Carrasco, J. A. (2013). Mode choice of university students commuting to school and the role of active travel. Journal of Transport Geography, 31, 132-142. (https://doi.org/10.1016/j.jtrangeo.2013.06.008)

Examples

 data(mc_commute_wide)
 summary(mc_commute_wide$choice)
#> Cycle  Walk   HSR   Car 
#>    47   711   336   281