McMaster Modality Data Set.

A dataset containing information about students, staff, and faculty commuting to McMaster University. Data were collected by means of a travel survey in the fall of 2010. Respondents were asked about their mode of travel to McMaster University, in Hamilton, Canada. They were also asked about the modes available to them. The data set also contains relevant attributes about the respondents. The format of the table is wide, with each row representing an individual respondent. These variables were augmented with information from the 2011 Census of Canada and land use information from local sources.

Usage

data(mc_modality)

Format

A data frame with 4,146 rows and 37 variables:

id

Unique identifier for respondents

choice

Mode of transportation chosen: Active Travel (Cycle/Walk), HSR (local transit), Car, GO (regional transit)

LAT

Latitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

LONG

Longitude of the place or residence, geocoded at the nearest major intersection or centroid of 6-digit postal area

shared_vehicle

Indicator variable for having individual access to a private car (i.e., car shared with other household members): Yes/No

bicycle

Indicator variable for owning a bicycle: Yes/No

gender

Woman/Man

age

Age in years

status

Indicator variable for status at McMaster: Staff or Faculty/Student

care_giver

Indicator variable for being primary care giver for any minors in the household: Yes/No

modality

Indicator variable for number of available modes: One Mode/Two Modes/Three Modes/Four Modes

Boring

Response to statement "Traveling is boring" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Buses_Comfortable

Response to statement "The buses I usually travel in are comfortable" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Clean_Fuel

Response to statement "To improve the environment, I am willing to pay a little more to use a hybrid, electric or other clean fuel vehicle" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Limit_Driving

Response to statement "I limit my auto travel to help improve congestion and the environment" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Productive_Time

Response to statement "I use my commute time productively" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Safe_Cycle

Response to statement "I feel safe and secure to cycle around town" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Shops_Services

Response to statement "Having shops and services within walking distance of my home is important to me" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Someone_Else

Response to statement "Usually, I'd rather have someone else do the driving" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Stuck_Traffic

Response to statement "Getting stuck in traffic doesn't bother me too much" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Transition

Response to statement "My commute trip is a useful transition between home and school" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Travel_Alone

Response to statement "I like traveling alone" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Travel_Tiring

Response to statement "Traveling is generally tiring for me" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Wasted_Time

Response to statement "Traveling is generally tiring for me" as factor: Strongly Disagree, Disagree, Neutral, Agree, Strongly Agree

Rate_Immigrant

Proportion of inmigrants to total population in Dissemination Area

Rate_Non_Canadian

Proportion of non-Canadian residents to total population in Dissemination Area

Rate_Labor

Proportion of residents in Dissemination Area who are in the labor force

Rate_Unemployment

Unemployment rate in Dissemination Area

Rate_Married

Proportion of families headed by a married couple to total population in Dissemination Area

Mean_Age

Mean age of residents in Dissemination Area in years

Mean_Children

Mean number of children per family in Dissemination Area

Median_HH_Income

Median family income in Dissemination Area in $10,000

LITA

Transit access index – combination of capacity, frequency, and coverage of service (Wiley et al., 2011)

LUM

Entropy-based land use mix index in Dissemination Area: lower values mean more homogenous land uses, higher values greater mix of uses

MAC_DISTANCE

Network distance from place of residence to McMaster University in km

POPULATION_DENSITY

Population density in Dissemination Area in people/sq.km

SF_P_RATIO

Ratio of building footprint to parcel area in Dissemination Area: a measure of density of development

Source

Lavery, T., P\'aez, A., & Kanaroglou, P.S. (2013). Driving out of choices: An investigation of transport modality in a university sample. Transportation Research Part A: Policy and Practice, 57:37-46 (https://doi.org/10.1016/j.tra.2013.09.010)

Examples

 data(mc_modality)
 summary(mc_modality)
#>        id                      choice          LAT             LONG       
#>  Min.   :1.170e+09   Active Travel:1686   Min.   :43.12   Min.   :-80.18  
#>  1st Qu.:1.173e+09   HSR          :1364   1st Qu.:43.24   1st Qu.:-79.93  
#>  Median :1.175e+09   Car          :1056   Median :43.25   Median :-79.91  
#>  Mean   :1.186e+09   GO           :  40   Mean   :43.25   Mean   :-79.90  
#>  3rd Qu.:1.194e+09                        3rd Qu.:43.26   3rd Qu.:-79.88  
#>  Max.   :1.241e+09                        Max.   :43.40   Max.   :-79.63  
#>  shared_vehicle bicycle      gender          age                     status    
#>  Yes :1099      No :1974   Man  :1573   Min.   :17.00   Staff or Faculty: 896  
#>  NA's:3047      Yes:2172   Woman:2573   1st Qu.:20.00   Student         :3250  
#>                                         Median :21.00                          
#>                                         Mean   :26.26                          
#>                                         3rd Qu.:28.00                          
#>                                         Max.   :71.00                          
#>  care_giver        modality                  Boring    
#>  No :3820   Four Modes : 137   STRONGLY DISAGREE: 444  
#>  Yes: 326   One Mode   : 888   DISAGREE         :1388  
#>             Three Modes:1110   NEUTRAL          :1342  
#>             Two Modes  :2011   AGREE            : 793  
#>                                STRONGLY AGREE   : 179  
#>                                                        
#>          Buses_Comfortable             Clean_Fuel             Limit_Driving 
#>  STRONGLY DISAGREE: 235    STRONGLY DISAGREE: 218   STRONGLY DISAGREE: 176  
#>  DISAGREE         : 745    DISAGREE         : 622   DISAGREE         : 861  
#>  NEUTRAL          :1432    NEUTRAL          :1151   NEUTRAL          :1209  
#>  AGREE            :1624    AGREE            :1624   AGREE            :1328  
#>  STRONGLY AGREE   : 110    STRONGLY AGREE   : 531   STRONGLY AGREE   : 572  
#>                                                                             
#>           Productive_Time             Safe_Cycle             Shops_Services
#>  STRONGLY DISAGREE: 372   STRONGLY DISAGREE: 364   STRONGLY DISAGREE:  35  
#>  DISAGREE         :1213   DISAGREE         : 982   DISAGREE         : 179  
#>  NEUTRAL          :1279   NEUTRAL          :1128   NEUTRAL          : 445  
#>  AGREE            :1066   AGREE            :1347   AGREE            :1974  
#>  STRONGLY AGREE   : 216   STRONGLY AGREE   : 325   STRONGLY AGREE   :1513  
#>                                                                            
#>             Someone_Else            Stuck_Traffic              Transition  
#>  STRONGLY DISAGREE: 434   STRONGLY DISAGREE:1239   STRONGLY DISAGREE: 263  
#>  DISAGREE         :1038   DISAGREE         :1627   DISAGREE         : 874  
#>  NEUTRAL          : 996   NEUTRAL          : 685   NEUTRAL          :1372  
#>  AGREE            :1191   AGREE            : 529   AGREE            :1348  
#>  STRONGLY AGREE   : 487   STRONGLY AGREE   :  66   STRONGLY AGREE   : 289  
#>                                                                            
#>             Travel_Alone            Travel_Tiring             Wasted_Time  
#>  STRONGLY DISAGREE: 224   STRONGLY DISAGREE: 177   STRONGLY DISAGREE: 206  
#>  DISAGREE         :1004   DISAGREE         :1385   DISAGREE         :1197  
#>  NEUTRAL          :1593   NEUTRAL          :1203   NEUTRAL          :1009  
#>  AGREE            :1129   AGREE            :1169   AGREE            :1323  
#>  STRONGLY AGREE   : 196   STRONGLY AGREE   : 212   STRONGLY AGREE   : 411  
#>                                                                            
#>  Rate_Immigrant   Rate_Non_Canadian   Rate_Labor     Rate_Unemployment
#>  Min.   :0.0000   Min.   :0.00000   Min.   :0.0000   Min.   :0.00000  
#>  1st Qu.:0.1887   1st Qu.:0.02326   1st Qu.:0.6090   1st Qu.:0.03200  
#>  Median :0.2660   Median :0.06021   Median :0.6620   Median :0.05900  
#>  Mean   :0.2715   Mean   :0.07959   Mean   :0.6602   Mean   :0.06559  
#>  3rd Qu.:0.3411   3rd Qu.:0.11538   3rd Qu.:0.7140   3rd Qu.:0.09400  
#>  Max.   :0.6723   Max.   :0.43750   Max.   :0.9110   Max.   :0.36500  
#>   Rate_Married       Mean_Age     Mean_Children   Median_HH_Income
#>  Min.   :0.0000   Min.   : 0.00   Min.   :0.000   Min.   : 0.000  
#>  1st Qu.:0.3663   1st Qu.:36.36   1st Qu.:0.900   1st Qu.: 4.577  
#>  Median :0.4370   Median :39.47   Median :1.100   Median : 5.989  
#>  Mean   :0.4513   Mean   :39.21   Mean   :1.078   Mean   : 6.248  
#>  3rd Qu.:0.5500   3rd Qu.:42.20   3rd Qu.:1.300   3rd Qu.: 7.651  
#>  Max.   :0.7453   Max.   :70.53   Max.   :2.500   Max.   :15.896  
#>       LITA           LUM          MAC_DISTANCE     POPULATION_DENSITY
#>  Min.   :1.00   Min.   :0.0000   Min.   : 0.2939   Min.   :  0.00    
#>  1st Qu.:3.00   1st Qu.:0.2443   1st Qu.: 1.2763   1st Qu.: 12.51    
#>  Median :4.00   Median :0.4480   Median : 3.4619   Median : 26.56    
#>  Mean   :3.76   Mean   :0.4002   Mean   : 5.5738   Mean   : 38.79    
#>  3rd Qu.:5.00   3rd Qu.:0.5837   3rd Qu.: 8.7451   3rd Qu.: 44.53    
#>  Max.   :5.00   Max.   :0.9130   Max.   :29.7833   Max.   :811.35    
#>    SF_P_RATIO    
#>  Min.   :0.0000  
#>  1st Qu.:0.1595  
#>  Median :0.2682  
#>  Mean   :0.2466  
#>  3rd Qu.:0.3160  
#>  Max.   :0.6753